Fixed timings, indicators and trimmed services.

+++ bootstrap.py
@@ -10,34 +10,29 @@
 from cryptocurrency.authentication import Cryptocurrency_authenticator
 from cryptocurrency.exchange import Cryptocurrency_exchange
 from cryptocurrency.conversion_table import get_conversion_table, get_new_tickers
-from cryptocurrency.bootstrapping import bootstrap_loggers
-
+from cryptocurrency.bootstrap import bootstrap_loggers
 import os
 import shutil
 
 def main():
+    as_pair = False
+    directory = 'crypto_logs' 
+    if os.path.exists(directory):
+        shutil.rmtree(directory)
+    os.mkdir(directory)
     authenticator = Cryptocurrency_authenticator(use_keys=False, testnet=False)
     client = authenticator.spot_client
-    exchange = Cryptocurrency_exchange(client=client, directory='crypto_logs')
+    exchange = Cryptocurrency_exchange(client=client, directory=directory)
     exchange_info = exchange.info
-    conversion_table = get_conversion_table(client=client, exchange_info=exchange_info, 
-                                            as_pair=True)
+    conversion_table = get_conversion_table(client=client, exchange_info=exchange_info, as_pair=True)
     assets = get_new_tickers(conversion_table=conversion_table)
-
-    directory = 'crypto_logs'
-    shutil.rmtree(directory)
-    os.mkdir(directory)
-
-    #pairs = bootstrap_loggers(client=client, assets=assets, 
-    #                          intervals=['1d', '7d', '30d'], pairs=None, 
-    #                          download_interval='1d', period=1000, second_period=None)
-    #pairs = bootstrap_loggers(client=client, assets=assets, 
-    #                          intervals=['1h', '2h', '4h', '12h'], pairs=pairs, 
-    #                          download_interval='1h', period=1000, second_period=None)
-    pairs = {}
-    pairs = bootstrap_loggers(client=client, assets=assets, 
-                              intervals=['1min', '5min', '15min', '30min'], pairs=pairs, 
-                              download_interval='1m', period=2880, second_period=60)
+    pairs = bootstrap_loggers(client=client, assets=assets, pairs={}, 
+                              download_interval='1d', exchange_info=exchange_info, as_pair=as_pair)
+    pairs = bootstrap_loggers(client=client, assets=assets, pairs=pairs, 
+                              download_interval='1h', exchange_info=exchange_info, as_pair=as_pair)
+    pairs = bootstrap_loggers(client=client, assets=assets, additional_intervals=['30min'], 
+                              upsampled_intervals=['5s', '15s'], pairs=pairs, 
+                              download_interval='1m', exchange_info=exchange_info, as_pair=as_pair)
     print('Bootstrapping done!')
 
 if __name__ == '__main__':
